+++ src/main/java/com/news/model/Organization.java
@@ -1,4 +1,4 @@
-package com.trip.butler.model;
+package com.news.model;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -7,13 +7,14 @@ import org.hibernate.annotations.OnDeleteAction;
 
 import javax.persistence.*;
 
+
 @Entity
 @NoArgsConstructor
 @Getter
-public class Person {
+public class Organization {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name= "person_id")
+    @Column(name= "organization_id")
     private Long id;
 
     @ManyToOne
@@ -21,5 +22,5 @@ public class Person {
     @OnDelete(action = OnDeleteAction.CASCADE)
     private News news;
 
-    private String person;
+    private String organization;
 }
